governance: add community manager role
Feel free to adjust wording as needed, this is just an example.

This is a good way to empower PMs other folks to be able to help out with meta things around the project. We explicitly list the the role in governance so that the people listed here end up in the main cncf maintainers.csv.

Being in the csv file is important because it is a check the CNCF uses to ensure maintainers are requesting the resources. And for CFP submissions, "did one of the maintainers submit this talk or was it someone else?", etc.

Signed-off-by: Laura Santamaria <nimbinatus@users.noreply.github.com>